Added data settings views for both skin temperature and step count.

+++ app/src/main/java/lab/star/surf_iot2015/sensor/SkinTempSensor.java
@@ -31,7 +31,7 @@ public class SkinTempSensor extends Sensor {
             public void onBandSkinTemperatureChanged
                     (BandSkinTemperatureEvent bandSkinTemperatureEvent) {
 
-                String valAsString = Double.toString(bandSkinTemperatureEvent.getTemperature());
+                String valAsString = Float.toString(bandSkinTemperatureEvent.getTemperature());
                 data.addEntry(currentTimeMillis(), valAsString);
 
                 for (SensorServiceCallback callback : callbacks){
